pyproject.toml
[tool.poetry]
name = "streaming-<NAME>"
version = "0.0.1"
description = ""
authors = ["Team_Phoenix <Team_Phoenix@alianza.com>"]
packages = [
    { include = "lambdas" },
]

[[tool.poetry.source]]
url = "https://alianza.jfrog.io/artifactory/api/pypi/alianza-pypi/simple"
name = "alianza-internal"
priority = "supplemental"

[[tool.poetry.source]]
url = "https://alianza.jfrog.io/alianza/api/pypi/alianza-snapshot-pypi/simple"
name = "alianza-snapshot"
priority = "supplemental"

[tool.poetry.dependencies]
python = ">=3.11.6,<4.0"
pytz = "2021.3"
streaming_lib = "0.0.25"
urllib3 = "<2"

[tool.poetry.group.dev.dependencies]
boto3 = { version = "*"}
black = "==23.3.0"
isort = "==5.12.0"
mypy = "==1.2.0"
mypy-extensions = "==1.0.0"
prospector = {version = "==1.9.0", extras = ["with_everything"]}
pylint = "==2.17.4"
pylint-celery = "==0.3"
pylint-django = "==2.5.3"
pylint-flask = "==0.6"
pylint-plugin-utils = "==0.7"
pyroma = "==4.2"
vulture = "==2.7"
pytest = "==7.3.1"
pytest-cov = "==4.0.0"
types-pytz = "==2022.1.2"
data_tests_lib = "==0.2.36"
allure-behave = "==2.13.2"
behave = "==1.2.6"
bs4 = "==0.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
